Add array support to the plist parser
The SysInfoExtended file on the ipod used to contain the artwork formats in
<array> tags containing <key> tags as well as the value. This is not valid
according to the plist DTD, and this was fixed with the 4th generation Nanos
(but not with the 120GB ipod classic).
libgpod was written using invalid (wrt to the XML DTD) files, and thus handled
<array> elements as if they were <dict> elements (because of the <key> tags).
This was wrong, and prevented it from parsing the fixed SysInfoExtended file
on the new Nanos. This changeset implements support for <array> tags (ignoring
tags like <key> in the process so as not to fail with older SysInfoExtended
files)

git-svn-id: https://gtkpod.svn.sourceforge.net/svnroot/gtkpod/libgpod/trunk@2135 f01d2545-417e-4e96-918e-98f8d0dbbcb6