Write album IDs and MHLAs to the database
Recent iTunes write a mhsd type 4 containing the list of all albums, this
commit adds an album id member to Itdb_Track and adds the code to generate
that ID and to write mhla/mhia headers to the itunesdb. Not sure what the
ipod uses that information for, but hopefully this will help to write
iTunesDB that the iphone likes, or to improve responsiveness of the album
list when artwork is involved.

git-svn-id: https://gtkpod.svn.sourceforge.net/svnroot/gtkpod/libgpod/trunk@2232 f01d2545-417e-4e96-918e-98f8d0dbbcb6