refactor reading of mhod strings and make sure they are valid UTF-8
The previous code didn't check at all if the strings read from the
iPod were valid UTF-8 strings which could cause problems later on.
It's also helpful in case we feed bogus data (for example an UTF-16
string with a bad endianness) to g_utf16_to_utf8 since in this case,
it returns bogus data too.