mhsd5 does not containt all smart playlists
It was assumed that the mhsd type 5 contained smart playlists
and the code was modified accordingly. However, it seems it's not
the case and that mhsd type 5 contains smart playlists referring
to your purchased items (tv shows, music, ringtones, ...). Revert
the previous changes to fix writing of podcasts/playlists/smart
playlists and for now, write empty mhsd5