split iphone callout out of the main hal callout
Doing the same things as the callout for regular iPods is scary since
the only way to browse an iPhone filesystem is through the (userland)
afc protocol, which might not be already set up when we are in a hal
callout. The most important things provided by the hal callout are
1. the SysInfoExtended file
2. the iPod name
3. the iPod model (to get a nice icon)

Since 2. and 3. are already done in another way by the GVFS AFC backend,
let's just do 1. in the iPhone hal callout, makes things less scary and
thus more maintainable.