make udev podsleuth-like information optional
The HAL callout exports various information about the ipod to the HAL
device tree to mimic what Podsleuth does. However, this kind of
information isn't something that belongs here and is one of the reasons
why HAL is being phased out in favour of udev + various specialized
daemon. Hence we don't want necessarily want to export it in the udev
callout, so make it optional.