[udev] make udev rule more strict
The nano5g has its block device and its partition which are both
identified by udev as containing a valid filesystem. However, only
the partition has an actual filesystem on it, not the block device.
Make sure the udev rule is only ran on the nano5g partition and not
on the block device.