rework itdb_create_directories
To properly create the ipod directory structure, this function needs to
differentiate between shuffles, iphone-like devices, ... The code used
to make some checks on ItdbIpodInfo::model but the checks weren't
exhaustive and only matched older shuffles and iphone-like devices.
Use the functions provided by itdb_device.c for that so that the
check are accurate.