Fixed 32 bit playlist writing
This patch changes the type of id from gulong to guint64, which is
what the dbid is stored as. This fixes the comparison between the
dbid between tracks for 32 bit systems and allows playlists to
be written. Many thanks to Stefan Ziel who's patch ended up
pointing out the problem.