Revert "Fix for creation of blank playlist on 80GB ipod video"
This reverts commit 1252bc0e18c396b2cb5e1bd5369f17403ca181aa.

On an iPod Video with 1.2.1 firmware:
-  With 1252bc0, no songs, no videos, no playlists at
all (except the On-the-go playlist)
-  Before 1252bc0, songs/videos/playlists are visible, but
there is an additional unnamed blank music playlist.

With latest firmware, none of these issues show up. The blank playlist
bug is better than the empty ipod bug, so let's revert to the earlier
behaviour and if someone encounters the blank playlist issue, we just
have to remember to tell him to update his firmware. For the record,
it seems only writing 4 mhsds on an iPod Video with firmware 1.2.1
avoids these 2 issues too.