fix a pretty bad bug accessing the g_lists from managed code
If managed land created a GLib.List from the native pointer and then we
did something to cause an item to be added/removed from the native list,
the managed copy would be out of sync and would be pointing at
just-freed memory. For example if the first item in the glist was
removed in native code, the managed copy would still think it was there.

We work around this issue by instantiating a GLib.List and letting it
die as soon as we use it once. We don't store it and re-use it.

Signed-off-by: Nathaniel McCallum <nathaniel@natemccallum.com>