Improve directory creation in itdb_init_ipod
For the creation of the iPod_Control directory, we first attempt
to reuse an existing directory using itdb_get_control_dir. This feels
safer than hardcoding the directory names as the old code does,
especially since the iPods seem to automatically recreate the
iPod_Control directory when it's deleted, The old code is still used
as a fallback for at least the iPod Nano 5th gen.