MediaType is a bitfield
1e719585f6 added new members to Itdb_Mediatype, but these members
are trivial binary OR of existing Itdb_MediaType elements. Media types
are actually bitfields, but this was never mentioned explicitly
anywhere. Improve Itdb_MediaType definition and documentation
to make that more obvious. This also fixes the few places where libgpod
manipulates media types to take into account the fact that it's a
bitfield.