[sqlite] Don't abort after the first copying error
After my recent changes, the code copying the temporary sqlite files
to the device would abort after the first copy error that happens.
Change it so that it tries to copy as many files as possible and
report the last error that occurred to be as close as the previous
behaviour as possible.