remove recent playcount file after db writing
Currently, when parsing an iPod database, libgpod reads the iPod
file containing information about the recent playcounts and merges
it with the database. However, it doesn't automatically remove this
file, and gives no way for applications to do it, as a result,
playcounts will be merged several times if itdb_parse is called
several time on an iPod. To avoid that, once the database has been
successfully written to disk, we make sure to remove the recent
playcount files.