attempt to bring sanity in shuffle generations
libgpod has been naming the 2.5g iPod Shuffle (which is physically
identical to the 2g iPod Shuffle) a 3rd generation iPod Shuffle
while for the rest of the world, the 3rd generation iPod Shuffle is
the button-less bar with voice-over.

Change libgpod to use the same convention as the rest of the world,
it's a little bit less accurate since now the 2g and 2.5g models
can't be differentiated, but the less confusing naming should be
better in the long run.  It's also technically an ABI breakage, but
I don't think anyone relies at all on the generation for an iPod
model not changing.