add more error reporting to sqlite generation
Someone reported an issue with gtkpod reporting "failed to write
sqlite DB" errors when trying to write to his Nano 5g. It was
finally tracked down to a missing HashInfo file which made
Locations.itdb.cbk generation fail. Add better error reporting
to that part of the code so that things are easier to track down
next time :)